requirements.txt
neo4j==5.9.0
flask==2.3.2
gunicorn==20.1.0
python-decouple==3.8
-i https://pypi.org/simple
fuzzywuzzy==0.18.0
isodate==0.6.1
numpy==1.24.3 ; python_version >= '3.10'
pyparsing==3.0.9 ; python_full_version >= '3.6.8'
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
rdflib==6.3.2

colorama==0.4.6 ; platform_system == 'Windows'
levenshtein==0.21.1 ; python_version >= '3.6'
pandas==2.0.2
python-levenshtein==0.21.1
pytz==2023.3
rapidfuzz==3.1.1 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
tqdm==4.65.0
tzdata==2023.3 ; python_version >= '2'